requirements.txt
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
grpcio==1.80.0
idna==3.4
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
protobuf==7.34.1
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.28.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.2.0+cpu
torchvision==0.17.0+cpu
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.13
Werkzeug==3.1.8
